requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# ML
scikit-learn==1.5.1
xgboost==3.0.5
lightgbm==4.6.0
catboost==1.2.8
imbalanced-learn==0.14.1
shap==0.49.1
optuna==4.6.0

# Deep Learning / LLM
torch==2.3.1
transformers==4.57.5
accelerate==0.31.0
sentence-transformers==5.2.0
sentencepiece==0.2.0
huggingface-hub==0.36.0
tokenizers==0.22.2
safetensors==0.4.3
bitsandbytes==0.43.1

# Vector DB
faiss-cpu==1.8.0.post1
chromadb==0.4.22

# Backend
fastapi==0.111.1
uvicorn==0.30.1
starlette==0.37.2
pydantic==2.8.0
orjson==3.10.6

# UI
streamlit==1.31.1

# Google GenAI (protobuf-safe)
protobuf==4.25.3
google-generativeai==0.8.6
google-ai-generativelanguage==0.6.15
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
grpcio-status==1.62.2
proto-plus==1.24.0
uritemplate==4.1.1

# Utilities
tqdm==4.66.4
rich==13.7.1
psutil==7.2.1
packaging==23.2
typing-extensions==4.15.0
